Fixed damageVehicleAtWorldPosition. Removed need for vehicleInitialOffset
Since 0,0,0 might not exist, it now uses a base voxel system. for what to pass into getVehiclePos. If 0,0,0 exists, its 0,0,0. If it doesnt exist, the position is the closest component. This also accidentally removed the need for the vehicle initial offsets for sub bodys, since the baseVoxelPos contains the voxel position, it automatically applys the offset to the vehicle position.