Even more optimizations - Read description
Increased collision checking performance by about 80%.

Think this is as much as i can optimize it other than super micro-optimizing everything. Main issue is matrix.multiply, apparently that is a super expensive function to use alot. Next step is going to be just avoiding more of the collision checks in the first place, like disabling collision with likely bombs