Removed BBOX Temp calls so i can compare a new matrix.multiply function performance
I was thinking about if we had how matrix.multiply works internally we could make it faster since thats the bottleneck. Apparently there is already a matrix.multiply which is faster. This may not be a working version, havent ran in-game yet but need to sync code since switching computers